feat: rota all com paginação dos post

+++ src/repositories/posts.ts
@@ -9,18 +9,39 @@ export default class Repository {
         this.prisma = new PrismaClient();
     }
 
-    public async delete(id: string): Promise<boolean> {
-        try {
-            await this.prisma.post.delete({
-                where: { id },
-                
-            });
-    
-            return true;
-        } catch(error) {
-            return false;
+    public async allPaged(index: number, size: number): Promise<Post[]> {
+        const postsFound = await this.prisma.post.findMany({
+            skip: size * index,
+            take: size
+        });
+
+        const posts: Post[] = [];
+
+        for(const post of postsFound) {
+            posts.push(new Post(
+                post.id,
+                post.image,
+                post.title,
+                post.description
+            ));
         }
 
+        return posts;
+    }
+
+    public async delete(id: string): Promise<boolean> {
+        const postFound = await this.prisma.post.findUnique({ where: { id }})
+
+        if (!postFound) {
+            return false;
+        }
+        
+        await this.prisma.post.delete({
+            where: { id },
+            
+        });
+
+        return true;
     }
 
     public async findById(id: string): Promise<Post | null> {
